+++ spark/v4.0/spark/src/main/java/org/apache/iceberg/spark/SparkSessionCatalog.java
@@ -464,13 +464,6 @@ public class SparkSessionCatalog<
     }
 
     if (null != asViewCatalog) {
-      // Check if the identifier is already taken in the underlying session catalog.
-      // The session catalog (V2SessionCatalog) is the authority on the namespace.
-      // tableExists() checks for both tables and views (V1 and V2).
-      if (getSessionCatalog().tableExists(viewInfo.ident())) {
-        throw new ViewAlreadyExistsException(viewInfo.ident());
-      }
-
       return asViewCatalog.createView(viewInfo);
     } else if (isViewCatalog()) {
       return getSessionCatalog().createView(viewInfo);
